@@ -62,7 +62,7 @@
     "README.md",
     "CHANGES.md",
     "lib",
-    "seneca.js"
+    "seneca.*"
   ],
   "engines": {
     "node": ">=14"
@@ -87,25 +87,25 @@
     "eraro": "^2.1.0",
     "fast-safe-stringify": "^2.1.1",
     "gate-executor": "^3.1.1",
-    "gubu": "^3.1.0",
+    "gubu": "^5.0.1",
     "jsonic": "^1.0.1",
     "lodash.defaultsdeep": "^4.6.1",
     "lodash.flatten": "^4.4.0",
     "lodash.uniq": "^4.5.0",
-    "minimist": "^1.2.6",
-    "nid": "^2.0.0",
-    "norma": "^2.0.2",
+    "minimist": "^1.2.8",
+    "nid": "^2.0.1",
+    "norma": "^3.0.0",
     "ordu": "^2.2.0",
-    "patrun": "^7.2.4",
-    "qs": "^6.10.5",
+    "patrun": "^7.2.5",
+    "qs": "^6.11.2",
     "rolling-stats": "^0.2.1",
     "use-plugin": "^10.0.0"
   },
   "devDependencies": {
-    "@hapi/code": "^9.0.1",
-    "@hapi/lab": "^25.0.1",
+    "@hapi/code": "^9.0.3",
+    "@hapi/lab": "^25.1.3",
     "@seneca/test-plugin": "0.1.0",
-    "@types/node": "^18.0.0",
+    "@types/node": "^20.3.2",
     "async": "^3.2.4",
     "bench": "^0.3.6",
     "coveralls": "^3.1.1",
@@ -113,11 +113,11 @@
     "handlebars": "^4.7.7",
     "lab-transform-typescript": "^3.0.1",
     "lolex": "^6.0.0",
-    "prettier": "^2.7.1",
-    "seneca-entity": "^17.0.1",
+    "prettier": "^2.8.8",
+    "seneca-entity": "^24.0.0",
     "seneca-error-test": "^0.2.2",
-    "seneca-promisify": "^2.4.1",
+    "seneca-promisify": "^3.6.0",
     "summary": "^2.1.0",
-    "typescript": "^4.7.4"
+    "typescript": "^5.1.6"
   }
 }
