@@ -1,4 +1,4 @@
-/* Copyright © 2010-2022 Richard Rodger and other contributors, MIT License. */
+/* Copyright © 2010-2023 Richard Rodger and other contributors, MIT License. */
 'use strict'
 
 
@@ -9,7 +9,7 @@ const Util = require('util')
 
 // External modules.
 const GateExecutor = require('gate-executor')
-const Jsonic = require('jsonic')
+const Jsonic = require('@jsonic/jsonic-next')
 const UsePlugin = require('use-plugin')
 import Nid from 'nid'
 import { Patrun, Gex } from 'patrun'
@@ -34,8 +34,8 @@ const Add = require('./lib/add')
 const Sub = require('./lib/sub')
 import { Prior } from './lib/prior'
 import { Plugin } from './lib/plugin'
-const { Inward } = require('./lib/inward')
-const { Outward } = require('./lib/outward')
+import { Inward } from './lib/inward'
+import { Outward } from './lib/outward'
 const { Legacy } = require('./lib/legacy')
 const { resolve_options } = require('./lib/options')
 const Package = require('./package.json')
@@ -43,6 +43,7 @@ const { Print } = require('./lib/print')
 const { addActions } = require('./lib/actions')
 const { transport } = require('./lib/transport')
 
+import Pkg from './package.json'
 
 // Internal data and utilities.
 const { error, deep } = Common
@@ -110,7 +111,10 @@ const option_defaults = {
 
       // Capture errors in actions and pass to callback (false throws uncaught).
       action: true,
-    }
+    },
+
+    // Custom function to identify thrown errors.
+    identify: (e: any) => e instanceof Error,
   },
 
   // Validate messages and options.
@@ -376,12 +380,12 @@ const option_defaults = {
 
 // Utility functions exposed by Seneca via `seneca.util`.
 const seneca_util = {
-  Eraro: Eraro,
+  Eraro,
 
-  Jsonic: Jsonic,
-  Nid: Nid,
-  Patrun: Patrun,
-  Gex: Gex,
+  Jsonic,
+  Nid,
+  Patrun,
+  Gex,
 
   clean: Common.clean,
   pattern: Common.pattern,
@@ -483,42 +487,51 @@ function init(seneca_options?: any, more_options?: any) {
   return seneca
 }
 
-module.exports = init
 
 // Expose Seneca prototype for easier monkey-patching
-module.exports.Seneca = Seneca
+init.Seneca = Seneca
 
-export default init
 
 // To reference builtin loggers when defining logging options.
-module.exports.loghandler = Legacy.loghandler
+init.loghandler = Legacy.loghandler
 
 // Makes require('seneca').use(...) work by creating an on-the-fly instance.
-module.exports.use = function top_use() {
+init.use = function top_use() {
   var argsarr = new Array(arguments.length)
   for (var l = 0; l < argsarr.length; ++l) {
     argsarr[l] = arguments[l]
   }
 
-  var instance = module.exports()
+  var instance = init()
 
   return instance.use.apply(instance, argsarr)
 }
 
 // Makes require('seneca').test() work.
-module.exports.test = function top_test() {
-  return module.exports().test(...arguments)
+init.test = function top_test() {
+  return init().test(...arguments)
 }
 
 // Makes require('seneca').quiet() work.
-module.exports.quiet = function top_quiet() {
-  return module.exports().quiet(...arguments)
+init.quiet = function top_quiet() {
+  return init().quiet(...arguments)
 }
 
-module.exports.util = seneca_util
-module.exports.valid = Gubu
+init.util = seneca_util
+init.valid = Gubu
 
-module.exports.test$ = { intern: intern }
+init.test$ = { intern: intern }
+
+
+type Instance = ReturnType<typeof make_seneca> & Record<string, any>
+
+export type { Instance }
+
+
+export default init
+
+module.exports = init
+
 
 // Create a new Seneca instance.
 function make_seneca(initial_opts?: any) {
@@ -545,7 +558,10 @@ function make_seneca(initial_opts?: any) {
   // These need to come from options as required during construction.
   private$.actrouter = start_opts.internal.actrouter || Patrun({ gex: true })
 
-  var soi_subrouter = start_opts.internal.subrouter || {}
+  private$.translationrouter =
+    start_opts.internal.translationrouter || Patrun({ gex: true })
+
+  const soi_subrouter = start_opts.internal.subrouter || {}
   private$.subrouter = {
     // Check for legacy inward router
     inward: soi_subrouter.inward || Patrun({ gex: true }),
@@ -574,10 +590,11 @@ function make_seneca(initial_opts?: any) {
   // Define public member variables.
   root$.start_time = Date.now()
   root$.context = {}
-  root$.version = Package.version
+  root$.version = Pkg.version
 
   // TODO: rename in 4.x as "args" terminology is legacy
   root$.fixedargs = {}
+  root$.fixedmeta = {}
 
   root$.flags = {
     closed: false,
@@ -770,6 +787,7 @@ function make_seneca(initial_opts?: any) {
     name: 'add',
     debug: !!start_opts.debug.ordu || !!start_opts.order.add.debug,
   })
+    .add(Add.task.translate)
     .add(Add.task.prepare)
     .add(Add.task.plugin)
     .add(Add.task.callpoint)
