@@ -57,7 +57,7 @@ const intern = (module.exports.intern = {
             }
             catch (e) {
                 if (opts.error.capture.action && true !== (e === null || e === void 0 ? void 0 : e.$$seneca_callback_error$$)) {
-                    const ex = (0, common_1.isError)(e) ? e : new Error((0, common_1.inspect)(e));
+                    const ex = opts.error.identify(e) ? e : new Error((0, common_1.inspect)(e));
                     intern.handle_reply(opts, meta, actctxt, actmsg, ex);
                     complete();
                 }
@@ -123,7 +123,7 @@ const intern = (module.exports.intern = {
         actctxt.actlog = intern.actlog;
         actctxt.errlog = intern.errlog;
         actctxt.error = common_1.error;
-        meta.error = (0, common_1.isError)(data.res);
+        meta.error = opts.error.identify(data.res);
         // A nasty edge case
         if (!meta.error && data.res === data.err) {
             data.err = null;
@@ -298,7 +298,7 @@ const intern = (module.exports.intern = {
         const reply = ctxt.reply;
         const meta = data.meta;
         const msg = data.msg;
-        let err = (0, common_1.isError)(thrown_obj)
+        let err = instance.options().error.identify(thrown_obj)
             ? thrown_obj
             : new Error((0, common_1.inspect)(thrown_obj));
         const opts = instance.options();
