@@ -4,7 +4,6 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.task = exports.api_add = void 0;
 const gubu_1 = require("gubu");
 const common_1 = require("./common");
-const Jsonic = require('jsonic');
 function api_add() {
     let ctx = {
         opts: this.options(),
@@ -28,9 +27,28 @@ function api_add() {
 }
 exports.api_add = api_add;
 const task = {
-    prepare(spec) {
+    translate(spec) {
         const args = spec.ctx.args;
         let raw_pattern = args.pattern;
+        let pattern = null;
+        if (false !== raw_pattern.translate$) {
+            // Must be exact match to ensure consistent translation
+            let translation = spec.ctx.private.translationrouter.find(raw_pattern);
+            if (translation) {
+                pattern = translation(raw_pattern);
+            }
+        }
+        return {
+            // TODO: simple op:set would be faster
+            op: 'merge',
+            out: {
+                pattern,
+            },
+        };
+    },
+    prepare(spec) {
+        const args = spec.ctx.args;
+        let raw_pattern = spec.data.pattern || args.pattern;
         let pattern = (0, common_1.clean)(raw_pattern);
         let action = args.action ||
             function default_action(_msg, done, meta) {
@@ -95,6 +113,7 @@ const task = {
         const actdef = spec.data.actdef;
         const pat = spec.data.pattern;
         const opts = spec.ctx.opts;
+        const Jsonic = spec.ctx.instance.util.Jsonic;
         actdef.sub = !!actdef.raw.sub$;
         actdef.client = !!actdef.raw.client$;
         // Deprecate a pattern by providing a string message using deprecate$ key.
@@ -128,8 +147,13 @@ const task = {
         const actdef = spec.data.actdef;
         const action = spec.data.action;
         const private$ = spec.ctx.private;
-        var action_name = null == action.name || '' === action.name ? 'action' : action.name;
-        actdef.id = action_name + '_' + private$.next_action_id();
+        const plugin = actdef.plugin || {};
+        const action_name = ((null == action.name || '' === action.name) ?
+            'action' : action.name);
+        actdef.id =
+            ((null == plugin.fullname || '' === plugin.fullname) ?
+                '' : plugin.fullname + '/') +
+                action_name + '/' + private$.next_action_id();
         actdef.name = action_name;
         actdef.func = action;
         if ('function' === typeof action.handle) {
@@ -182,6 +206,7 @@ const task = {
             },
         };
     },
+    // TODO: accept action.valid to define rules
     rules(spec) {
         const opts = spec.ctx.opts;
         const actdef = spec.data.actdef;
