@@ -1,19 +1,18 @@
 "use strict";
-/* Copyright © 2010-2022 Richard Rodger and other contributors, MIT License. */
+/* Copyright © 2010-2023 Richard Rodger and other contributors, MIT License. */
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.API = void 0;
-var Jsonic = require('jsonic');
-var Norma = require('norma');
-var Common = require('./common');
-var errlog = Common.make_standard_err_log_entry;
-var intern = {};
+const Norma = require('norma');
+const common_1 = require("./common");
+const errlog = common_1.make_standard_err_log_entry;
+const intern = {};
 function wrap(pin, actdef, wrapper) {
     var pinthis = this;
     wrapper = 'function' === typeof actdef ? actdef : wrapper;
     actdef = 'function' === typeof actdef ? {} : actdef;
     pin = Array.isArray(pin) ? pin : [pin];
-    Common.each(pin, function (p) {
-        Common.each(pinthis.list(p), function (actpattern) {
+    (0, common_1.each)(pin, function (p) {
+        (0, common_1.each)(pinthis.list(p), function (actpattern) {
             pinthis.add(actpattern, wrapper, actdef);
         });
     });
@@ -21,7 +20,7 @@ function wrap(pin, actdef, wrapper) {
 }
 function fix(patargs, msgargs, custom) {
     var self = this;
-    patargs = Jsonic(patargs || {});
+    patargs = self.util.Jsonic(patargs || {});
     var fix_delegate = self.delegate(patargs);
     // TODO: attach msgargs and custom to delegate.private$ in some way for debugging
     fix_delegate.add = function fix_add() {
@@ -64,7 +63,7 @@ function options(options, chain) {
     }
     // Update callpoint
     if (out_opts.debug.callpoint) {
-        private$.callpoint = Common.make_callpoint(out_opts.debug.callpoint);
+        private$.callpoint = (0, common_1.make_callpoint)(out_opts.debug.callpoint);
     }
     // DEPRECATED
     if (out_opts.legacy.logging) {
@@ -85,7 +84,7 @@ function close(callpoint) {
     return function api_close(done) {
         var seneca = this;
         if (false !== done && null == done) {
-            return Common.promiser(intern.close.bind(seneca, callpoint));
+            return (0, common_1.promiser)(intern.close.bind(seneca, callpoint));
         }
         return intern.close.call(seneca, callpoint, done);
     };
@@ -127,7 +126,7 @@ function error(first) {
             error = plugin.eraro.apply(this, arguments);
         }
         else {
-            error = Common.eraro.apply(this, arguments);
+            error = common_1.error.apply(this, arguments);
         }
         return error;
     }
@@ -199,7 +198,7 @@ function delegate(fixedargs, fixedmeta) {
         });
         strdesc =
             self.toString() +
-                (Object.keys(vfa).length ? '/' + Jsonic.stringify(vfa) : '');
+                (Object.keys(vfa).length ? '/' + (0, common_1.jsonic_stringify)(vfa) : '');
         return strdesc;
     }
     var delegate_fixedargs = opts.strict.fixedargs
@@ -330,10 +329,10 @@ function ping() {
         }),
     };
 }
-function translate(from_in, to_in, pick_in) {
-    var from = 'string' === typeof from_in ? Jsonic(from_in) : from_in;
-    var to = 'string' === typeof to_in ? Jsonic(to_in) : to_in;
-    var pick = {};
+function translate(from_in, to_in, pick_in, flags) {
+    const from = 'string' === typeof from_in ? this.util.Jsonic(from_in) : from_in;
+    const to = 'string' === typeof to_in ? this.util.Jsonic(to_in) : to_in;
+    let pick = {};
     if ('string' === typeof pick_in) {
         pick_in = pick_in.split(/\s*,\s*/);
     }
@@ -353,8 +352,8 @@ function translate(from_in, to_in, pick_in) {
     else {
         pick = null;
     }
-    this.add(from, function (msg, reply) {
-        var pick_msg;
+    let translate = function (msg) {
+        let pick_msg;
         if (pick) {
             pick_msg = {};
             Object.keys(pick).forEach(function (prop) {
@@ -364,11 +363,29 @@ function translate(from_in, to_in, pick_in) {
             });
         }
         else {
-            pick_msg = this.util.clean(msg);
+            pick_msg = (0, common_1.clean)(msg);
         }
-        var transmsg = Object.assign(pick_msg, to);
+        let transmsg = Object.assign(pick_msg, to);
+        for (let pn in transmsg) {
+            if (null == transmsg[pn]) {
+                delete transmsg[pn];
+            }
+        }
+        return transmsg;
+    };
+    this.private$.translationrouter.add(from, translate);
+    // console.log('TTT', translate({ c: 3, q: 9 }))
+    // console.log(this.private$.translationrouter + '')
+    let translation_action = function (msg, reply) {
+        let transmsg = translate(msg);
+        // console.log('ACT T', msg, transmsg)
         this.act(transmsg, reply);
+    };
+    Object.defineProperty(translation_action, 'name', {
+        value: 'translation__' + (0, common_1.jsonic_stringify)(from) + '__' + (0, common_1.jsonic_stringify)(to)
     });
+    from.translate$ = false;
+    this.add(from, translation_action);
     return this;
 }
 function gate() {
@@ -383,11 +400,11 @@ function list_plugins() {
     return Object.assign({}, this.private$.plugins);
 }
 function find_plugin(plugindesc, tag) {
-    var plugin_key = Common.make_plugin_key(plugindesc, tag);
+    var plugin_key = (0, common_1.make_plugin_key)(plugindesc, tag);
     return this.private$.plugins[plugin_key];
 }
 function has_plugin(plugindesc, tag) {
-    var plugin_key = Common.make_plugin_key(plugindesc, tag);
+    var plugin_key = (0, common_1.make_plugin_key)(plugindesc, tag);
     return !!this.private$.plugins[plugin_key];
 }
 function ignore_plugin(plugindesc, tag, ignore) {
@@ -395,7 +412,7 @@ function ignore_plugin(plugindesc, tag, ignore) {
         ignore = tag;
         tag = null;
     }
-    var plugin_key = Common.make_plugin_key(plugindesc, tag);
+    var plugin_key = (0, common_1.make_plugin_key)(plugindesc, tag);
     var resolved_ignore = (this.private$.ignore_plugins[plugin_key] =
         null == ignore ? true : !!ignore);
     this.log.info({
@@ -409,7 +426,7 @@ function ignore_plugin(plugindesc, tag, ignore) {
 // Find the action metadata for a given pattern, if it exists.
 function find(pattern, flags) {
     var seneca = this;
-    var pat = 'string' === typeof pattern ? Jsonic(pattern) : pattern;
+    var pat = 'string' === typeof pattern ? seneca.util.Jsonic(pattern) : pattern;
     pat = seneca.util.clean(pat);
     pat = pat || {};
     var actdef = seneca.private$.actrouter.find(pat, flags && flags.exact);
@@ -425,14 +442,8 @@ function has(pattern) {
 // List all actions that match the pattern.
 function list(pattern) {
     return this.private$.actrouter
-        .list(null == pattern ? {} : Jsonic(pattern))
+        .list(null == pattern ? {} : this.util.Jsonic(pattern))
         .map((x) => x.match);
-    /*
-    return _.map(
-      this.private$.actrouter.list(null == pattern ? {} : Jsonic(pattern)),
-      'match'
-    )
-    */
 }
 // Get the current status of the instance.
 function status(flags) {
@@ -505,13 +516,13 @@ function client(callpoint) {
         var opts = self.options().transport || {};
         var raw_config = intern.parse_config(argsarr);
         // pg: pin group
-        raw_config.pg = Common.pincanon(raw_config.pin || raw_config.pins);
+        raw_config.pg = (0, common_1.pincanon)(raw_config.pin || raw_config.pins);
         var config = intern.resolve_config(raw_config, opts);
-        config.id = config.id || Common.pattern(raw_config);
+        config.id = config.id || (0, common_1.pattern)(raw_config);
         var pins = config.pins ||
             (Array.isArray(config.pin) ? config.pin : [config.pin || '']);
         pins = pins.map((pin) => {
-            return 'string' === typeof pin ? Jsonic(pin) : pin;
+            return 'string' === typeof pin ? self.util.Jsonic(pin) : pin;
         });
         //var sd = Plugins.make_delegate(self, {
         // TODO: review - this feels like a hack
@@ -571,7 +582,7 @@ function client(callpoint) {
                 return sd.die(private$.error(err, 'transport_client', config));
             }
             if (null == liveclient) {
-                return sd.die(private$.error('transport_client_null', Common.clean(config)));
+                return sd.die(private$.error('transport_client_null', (0, common_1.clean)(config)));
             }
             sendclient = liveclient;
         });
@@ -667,7 +678,7 @@ intern.close = function (callpoint, done) {
         seneca.closed = true;
         seneca.flags.closed = true;
         // cleanup process event listeners
-        Common.each(options.system.close_signals, function (active, signal) {
+        (0, common_1.each)(options.system.close_signals, function (active, signal) {
             if (active) {
                 process.removeListener(signal, seneca.private$.exit_close);
             }
@@ -696,7 +707,7 @@ intern.close = function (callpoint, done) {
 };
 intern.fix_args =
     function (origargs, patargs, msgargs, custom) {
-        var args = Common.parsePattern(this, origargs, 'rest:.*', patargs);
+        var args = (0, common_1.parsePattern)(this, origargs, 'rest:.*', patargs);
         var fixargs = [args.pattern]
             .concat({
             fixed$: Object.assign({}, msgargs, args.pattern.fixed$),
