@@ -4,12 +4,12 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.TRACE_ACTION = exports.TRACE_SYNC = exports.TRACE_END = exports.TRACE_START = exports.TRACE_VERSION = exports.TRACE_TAG = exports.TRACE_INSTANCE = exports.TRACE_ID = exports.TRACE_PATTERN = exports.msgstr = exports.error = exports.inspect = exports.isError = exports.tagnid = exports.parsePattern = exports.print = exports.history = exports.make_trace_desc = exports.make_callpoint = exports.autoincr = exports.resolve_option = exports.make_standard_err_log_entry = exports.make_standard_act_log_entry = exports.makedie = exports.each = exports.deep = exports.clean = exports.noop = exports.pincanon = exports.pattern = exports.build_message = exports.parse_pattern = exports.parse_jsonic = exports.boolify = exports.make_plugin_key = exports.wrap_error = exports.stringify = exports.promiser = void 0;
+exports.TRACE_ACTION = exports.TRACE_SYNC = exports.TRACE_END = exports.TRACE_START = exports.TRACE_VERSION = exports.TRACE_TAG = exports.TRACE_INSTANCE = exports.TRACE_ID = exports.TRACE_PATTERN = exports.jsonic_stringify = exports.msgstr = exports.error = exports.inspect = exports.isError = exports.tagnid = exports.parsePattern = exports.print = exports.history = exports.make_trace_desc = exports.make_callpoint = exports.autoincr = exports.resolve_option = exports.make_standard_err_log_entry = exports.make_standard_act_log_entry = exports.makedie = exports.each = exports.deep = exports.clean = exports.noop = exports.pincanon = exports.pattern = exports.build_message = exports.parse_pattern = exports.parse_jsonic = exports.boolify = exports.make_plugin_key = exports.wrap_error = exports.stringify = exports.promiser = void 0;
 const util_1 = __importDefault(require("util"));
 const errors_1 = __importDefault(require("./errors"));
 const Stringify = require('fast-safe-stringify');
 const Eraro = require('eraro');
-const Jsonic = require('jsonic');
+const Jsonic = require('@jsonic/jsonic-next');
 const nid_1 = __importDefault(require("nid"));
 const Norma = require('norma');
 const DefaultsDeep = require('lodash.defaultsdeep');
@@ -100,19 +100,19 @@ function parse_jsonic(str, code) {
         return null == str ? null : Jsonic(str);
     }
     catch (e) {
-        let col = 1 === e.line ? e.column - 1 : e.column;
+        // let col = 1 === e.line ? e.column - 1 : e.column
         throw error(code, {
             argstr: str,
             syntax: e.message,
-            line: e.line,
-            col: col,
+            line: e.lineNumber,
+            col: e.columnNumber,
         });
     }
 }
 exports.parse_jsonic = parse_jsonic;
 // string args override object args
 // TODO: fix name
-function parse_pattern(instance, rawargs, normaspec, fixed) {
+function parse_pattern(_instance, rawargs, normaspec, fixed) {
     let args = Norma('{strargs:s? objargs:o? moreobjargs:o? ' + (normaspec || '') + '}', rawargs);
     // Precedence of arguments in add,act is left-to-right
     args.pattern = Object.assign({}, args.moreobjargs ? args.moreobjargs : null, args.objargs ? args.objargs : null, parse_jsonic(args.strargs, 'add_string_pattern_syntax'), fixed);
@@ -121,7 +121,7 @@ function parse_pattern(instance, rawargs, normaspec, fixed) {
 exports.parse_pattern = parse_pattern;
 const parsePattern = parse_pattern;
 exports.parsePattern = parsePattern;
-function build_message(instance, rawargs, normaspec, fixed) {
+function build_message(_instance, rawargs, normaspec, fixed) {
     let args = Norma('{strargs:s? objargs:o? moreobjargs:o? ' + (normaspec || '') + '}', rawargs);
     // Precedence of arguments in add,act is left-to-right
     args.msg = Object.assign({}, args.moreobjargs, args.objargs, parse_jsonic(args.strargs, 'msg_jsonic_syntax'), fixed);
@@ -282,7 +282,7 @@ function makedie(instance, ctxt) {
                 ', ' +
                 when.getTime() +
                 '\nLOG       :::  ' +
-                Jsonic.stringify(logdesc) +
+                jsonic_stringify(logdesc) +
                 '\nNODE      :::  ' +
                 process.version +
                 ', ' +
@@ -467,6 +467,95 @@ function msgstr(msg, len = 111) {
     return str;
 }
 exports.msgstr = msgstr;
+function jsonic_strify(val, opts, depth) {
+    depth++;
+    if (null == val)
+        return 'null';
+    var type = Object.prototype.toString.call(val).charAt(8);
+    if ('F' === type && !opts.showfunc)
+        return null;
+    // WARNING: output may not be jsonically parsable!
+    if (opts.custom) {
+        if (Object.prototype.hasOwnProperty.call(val, 'toString')) {
+            return val.toString();
+        }
+        else if (Object.prototype.hasOwnProperty.call(val, 'inspect')) {
+            return val.inspect();
+        }
+    }
+    var out, i = 0, j, k;
+    if ('N' === type) {
+        return isNaN(val) ? 'null' : val.toString();
+    }
+    else if ('O' === type) {
+        out = [];
+        if (depth <= opts.depth) {
+            j = 0;
+            for (let i in val) {
+                if (j >= opts.maxitems)
+                    break;
+                var pass = true;
+                for (k = 0; k < opts.exclude.length && pass; k++) {
+                    pass = !~i.indexOf(opts.exclude[k]);
+                }
+                pass = pass && !opts.omit[i];
+                let str = jsonic_strify(val[i], opts, depth);
+                if (null != str && pass) {
+                    var n = i.match(/^[a-zA-Z0-9_$]+$/) ? i : JSON.stringify(i);
+                    out.push(n + ':' + str);
+                    j++;
+                }
+            }
+        }
+        return '{' + out.join(',') + '}';
+    }
+    else if ('A' === type) {
+        out = [];
+        if (depth <= opts.depth) {
+            for (; i < val.length && i < opts.maxitems; i++) {
+                let str = jsonic_strify(val[i], opts, depth);
+                if (null != str) {
+                    out.push(str);
+                }
+            }
+        }
+        return '[' + out.join(',') + ']';
+    }
+    else {
+        var valstr = val.toString();
+        if (~" \"'\r\n\t,}]".indexOf(valstr[0]) ||
+            !~valstr.match(/,}]/) ||
+            ~" \r\n\t".indexOf(valstr[valstr.length - 1])) {
+            valstr = "'" + valstr.replace(/'/g, "\\'") + "'";
+        }
+        return valstr;
+    }
+}
+// Legacy Jsonic stringify
+function jsonic_stringify(val, callopts) {
+    try {
+        callopts = callopts || {};
+        var opts = {};
+        opts.showfunc = callopts.showfunc || callopts.f || false;
+        opts.custom = callopts.custom || callopts.c || false;
+        opts.depth = callopts.depth || callopts.d || 3;
+        opts.maxitems = callopts.maxitems || callopts.mi || 11;
+        opts.maxchars = callopts.maxchars || callopts.mc || 111;
+        opts.exclude = callopts.exclude || callopts.x || ['$'];
+        var omit = callopts.omit || callopts.o || [];
+        opts.omit = {};
+        for (var i = 0; i < omit.length; i++) {
+            opts.omit[omit[i]] = true;
+        }
+        var str = jsonic_strify(val, opts, 0);
+        str = null == str ? '' : str.substring(0, opts.maxchars);
+        return str;
+    }
+    catch (e) {
+        return 'ERROR: jsonic_stringify: ' + e + ' input was: ' + JSON.stringify(val);
+    }
+}
+exports.jsonic_stringify = jsonic_stringify;
 const TRACE_PATTERN = 0;
 exports.TRACE_PATTERN = TRACE_PATTERN;
 const TRACE_ID = 1;
