@@ -10,7 +10,7 @@ import Errors from './errors'
 
 const Stringify = require('fast-safe-stringify')
 const Eraro = require('eraro')
-const Jsonic = require('jsonic')
+const Jsonic = require('@jsonic/jsonic-next')
 import Nid from 'nid'
 const Norma = require('norma')
 const DefaultsDeep = require('lodash.defaultsdeep')
@@ -31,7 +31,7 @@ const error =
     }))
 
 
-function promiser(context: any, callback: any) {
+function promiser(context: any, callback?: any) {
   if ('function' === typeof context && null == callback) {
     callback = context
   } else {
@@ -122,12 +122,12 @@ function parse_jsonic(str: any, code: any) {
   try {
     return null == str ? null : Jsonic(str)
   } catch (e: any) {
-    let col = 1 === e.line ? e.column - 1 : e.column
+    // let col = 1 === e.line ? e.column - 1 : e.column
     throw error(code, {
       argstr: str,
       syntax: e.message,
-      line: e.line,
-      col: col,
+      line: e.lineNumber,
+      col: e.columnNumber,
     })
   }
 }
@@ -137,7 +137,7 @@ function parse_jsonic(str: any, code: any) {
 // TODO: fix name
 
 function parse_pattern(
-  instance: any,
+  _instance: any,
   rawargs: any,
   normaspec: any,
   fixed?: any
@@ -163,7 +163,7 @@ const parsePattern = parse_pattern
 
 
 function build_message(
-  instance: any,
+  _instance: any,
   rawargs: any,
   normaspec: any,
   fixed: any
@@ -370,7 +370,7 @@ function makedie(instance: any, ctxt: any) {
         ', ' +
         when.getTime() +
         '\nLOG       :::  ' +
-        Jsonic.stringify(logdesc) +
+        jsonic_stringify(logdesc) +
         '\nNODE      :::  ' +
         process.version +
         ', ' +
@@ -600,6 +600,109 @@ function msgstr(msg: any, len: number = 111): string {
 }
 
 
+function jsonic_strify(val: any, opts: any, depth: number) {
+  depth++
+  if (null == val) return 'null'
+
+  var type = Object.prototype.toString.call(val).charAt(8)
+  if ('F' === type && !opts.showfunc) return null
+
+  // WARNING: output may not be jsonically parsable!
+  if (opts.custom) {
+    if (Object.prototype.hasOwnProperty.call(val, 'toString')) {
+      return val.toString()
+    }
+    else if (Object.prototype.hasOwnProperty.call(val, 'inspect')) {
+      return val.inspect()
+    }
+  }
+
+
+  var out, i = 0, j, k
+
+  if ('N' === type) {
+    return isNaN(val) ? 'null' : val.toString()
+  }
+  else if ('O' === type) {
+    out = []
+    if (depth <= opts.depth) {
+      j = 0
+      for (let i in val) {
+        if (j >= opts.maxitems) break
+
+        var pass = true
+        for (k = 0; k < opts.exclude.length && pass; k++) {
+          pass = !~i.indexOf(opts.exclude[k])
+        }
+        pass = pass && !opts.omit[i]
+
+        let str: string = jsonic_strify(val[i], opts, depth)
+
+        if (null != str && pass) {
+          var n = i.match(/^[a-zA-Z0-9_$]+$/) ? i : JSON.stringify(i)
+          out.push(n + ':' + str)
+          j++
+        }
+      }
+    }
+    return '{' + out.join(',') + '}'
+  }
+  else if ('A' === type) {
+    out = []
+    if (depth <= opts.depth) {
+      for (; i < val.length && i < opts.maxitems; i++) {
+        let str: string = jsonic_strify(val[i], opts, depth)
+        if (null != str) {
+          out.push(str)
+        }
+      }
+    }
+    return '[' + out.join(',') + ']'
+  }
+  else {
+    var valstr = val.toString()
+
+    if (~" \"'\r\n\t,}]".indexOf(valstr[0]) ||
+      !~valstr.match(/,}]/) ||
+      ~" \r\n\t".indexOf(valstr[valstr.length - 1])) {
+      valstr = "'" + valstr.replace(/'/g, "\\'") + "'"
+    }
+
+    return valstr
+  }
+}
+
+
+
+// Legacy Jsonic stringify
+function jsonic_stringify(val: any, callopts?: any) {
+  try {
+    callopts = callopts || {}
+    var opts: any = {}
+
+    opts.showfunc = callopts.showfunc || callopts.f || false
+    opts.custom = callopts.custom || callopts.c || false
+    opts.depth = callopts.depth || callopts.d || 3
+    opts.maxitems = callopts.maxitems || callopts.mi || 11
+    opts.maxchars = callopts.maxchars || callopts.mc || 111
+    opts.exclude = callopts.exclude || callopts.x || ['$']
+    var omit = callopts.omit || callopts.o || []
+
+    opts.omit = {}
+    for (var i = 0; i < omit.length; i++) {
+      opts.omit[omit[i]] = true;
+    }
+
+    var str: string = jsonic_strify(val, opts, 0)
+    str = null == str ? '' : str.substring(0, opts.maxchars)
+    return str
+  }
+  catch (e) {
+    return 'ERROR: jsonic_stringify: ' + e + ' input was: ' + JSON.stringify(val)
+  }
+}
+
+
 const TRACE_PATTERN = 0
 const TRACE_ID = 1
 const TRACE_INSTANCE = 2
@@ -751,6 +854,7 @@ export {
   inspect,
   error,
   msgstr,
+  jsonic_stringify,
   TRACE_PATTERN,
   TRACE_ID,
   TRACE_INSTANCE,
