@@ -10,7 +10,7 @@ const errors_1 = __importDefault(require("./errors"));
 const Flatten = require('lodash.flatten');
 const Eraro = require('eraro');
 const Norma = require('norma');
-const Jsonic = require('jsonic');
+const Jsonic = require('@jsonic/jsonic-next');
 const Common = require('./common');
 const internals = {
     error: Eraro({
