@@ -316,8 +316,7 @@ function build_test_log(seneca, data) {
             let objstr = util_1.default.inspect(seneca.util.clean(obj))
                 .replace(/\s+/g, '')
                 .substring(0, datalen);
-            if (objstr.length <= 22 ||
-                (!obj.$$logged$$ && (!data.err || !data.err.$$logged$$))) {
+            if (objstr.length <= 22) {
                 logb.push(objstr);
                 if ('object' === typeof obj) {
                     obj.$$logged$$ = () => { };
@@ -336,16 +335,13 @@ function build_test_log(seneca, data) {
         if (data.data) {
             logb.push(data.data);
         }
-        if ('ERR' === data.case && data.err && !data.err.$$logged$$) {
+        if ('ERR' === data.case && data.err) {
             logb.push((data.err.code ? '\n\n' + data.err.code : '') +
                 '\n\n' +
                 data.err.stack +
                 '\n' +
                 data.caller +
                 '\n');
-            if ('object' === typeof data.err) {
-                data.err.$$logged$$ = () => { };
-            }
         }
     }
     else if ('add' === data.kind) {
@@ -377,11 +373,14 @@ function build_test_log(seneca, data) {
             'function' === typeof config.port ? '' : config.port,
         ].join(';'));
     }
-    else if (!data.$$logged$$) {
-        logb.push(util_1.default.inspect(data).replace(/\n/g, ' ').substring(0, datalen));
-        if ('object' === typeof data) {
-            data.$$logged$$ = () => { };
-        }
+    // Just plain data
+    else {
+        // TODO: use jsonic util
+        let datastr = util_1.default.inspect(seneca.util.clean(data.data || data))
+            .replace(/\s+/g, '');
+        // .substring(0, datalen)
+        logb.push(datastr.length <= datalen ?
+            datastr : datastr.substring(0, datalen) + '...');
     }
     if (data.did) {
         logb.push(data.did);
