@@ -8,7 +8,7 @@ exports.resolve_options = void 0;
 const fs_1 = __importDefault(require("fs"));
 const path_1 = __importDefault(require("path"));
 const Eraro = require('eraro');
-const Jsonic = require('jsonic');
+const Jsonic = require('@jsonic/jsonic-next');
 const Minimist = require('minimist');
 const { Gubu } = require('gubu');
 const Common = require('./common');
