@@ -1,4 +1,4 @@
-/* Copyright © 2020 Richard Rodger and other contributors, MIT License. */
+/* Copyright © 2020-2023 Richard Rodger and other contributors, MIT License. */
 /* $lab:coverage:off$ */
 'use strict';
 var __importDefault = (this && this.__importDefault) || function (mod) {
@@ -6,6 +6,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.Plugin = void 0;
+// TODO: replace `seneca-` prefix with `plugin-` for unnamed plugins
 const Uniq = require('lodash.uniq');
 const Eraro = require('eraro');
 const nua_1 = __importDefault(require("nua"));
@@ -26,6 +27,13 @@ function api_use(callpoint, opts) {
         tasks.args,
         tasks.load,
         tasks.normalize,
+        {
+            name: 'pre_options', exec: (spec) => {
+                if ('function' === typeof spec.data.plugin.define.preload) {
+                    return tasks.options(spec);
+                }
+            }
+        },
         tasks.preload,
         { name: 'pre_meta', exec: tasks.meta },
         { name: 'pre_legacy_extend', exec: tasks.legacy_extend },
@@ -118,6 +126,12 @@ function make_tasks() {
             let args = spec.data.plugin.args;
             let seneca = spec.ctx.seneca;
             let private$ = seneca.private$;
+            // Special cases for short plugin names
+            // TODO: plugin loading should check for @seneca and seneca first!
+            // 1. Avoid conflict with the OG request module!
+            if ('request' === args[0]) {
+                args[0] = '@seneca/request';
+            }
             // TODO: use-plugin needs better error message for malformed plugin desc
             let desc = private$.use.build_plugin_desc(...args);
             desc.callpoint = spec.ctx.callpoint;
@@ -360,24 +374,17 @@ function make_tasks() {
         },
         options: (spec) => {
             let plugin = spec.data.plugin;
-            let delegate = spec.data.delegate;
+            let delegate = spec.data.delegate ||
+                spec.ctx.seneca; // for preload
             let so = delegate.options();
             let fullname = plugin.fullname;
             let defaults = plugin.defaults;
-            let fullname_options = Object.assign({}, 
-            // DEPRECATED: remove in 4
-            so[fullname], so.plugin[fullname], 
-            // DEPRECATED: remove in 4
-            so[fullname + '$' + plugin.tag], so.plugin[fullname + '$' + plugin.tag]);
+            let fullname_options = Object.assign({}, so.plugin[fullname], so.plugin[fullname + '$' + plugin.tag]);
             let shortname = fullname !== plugin.name ? plugin.name : null;
             if (!shortname && fullname.indexOf('seneca-') === 0) {
                 shortname = fullname.substring('seneca-'.length);
             }
-            let shortname_options = Object.assign({}, 
-            // DEPRECATED: remove in 4
-            so[shortname], so.plugin[shortname], 
-            // DEPRECATED: remove in 4
-            so[shortname + '$' + plugin.tag], so.plugin[shortname + '$' + plugin.tag]);
+            let shortname_options = Object.assign({}, so.plugin[shortname], so.plugin[shortname + '$' + plugin.tag]);
             let base = {};
             // NOTE: plugin error codes are in their own namespaces
             // TODO: test this!!!
