@@ -7,12 +7,20 @@ const Common = require('./common');
 function api_sub() {
     const self = this;
     const subargs = Common.parsePattern(self, arguments, 'action:f');
-    const pattern = subargs.pattern;
+    const raw_pattern = subargs.pattern;
     const action = subargs.action;
-    let is_inward = !!pattern.in$;
-    let is_outward = !!pattern.out$;
+    let is_inward = !!raw_pattern.in$;
+    let is_outward = !!raw_pattern.out$;
     if (!is_outward) {
         is_inward = true; // Default if nothing specified
+    }
+    let pattern = raw_pattern;
+    if (false !== raw_pattern.translate$) {
+        // Must be exact match to ensure consistent translation
+        let translation = this.private$.translationrouter.find(raw_pattern);
+        if (translation) {
+            pattern = translation(raw_pattern);
+        }
     }
     const sub_pattern = self.util.clean(pattern);
     const routers = [
