@@ -10,7 +10,7 @@ const http_1 = __importDefault(require("http"));
 const Https = require('https');
 const Qs = require('qs');
 const Url = require('url');
-const Jsonic = require('jsonic');
+const Jsonic = require('@jsonic/jsonic-next');
 const Wreck = require('@hapi/wreck');
 const Common = require('./common');
 const legacy_1 = require("./legacy");
