@@ -1,4 +1,4 @@
-/* Copyright © 2010-2022 Richard Rodger and other contributors, MIT License. */
+/* Copyright © 2010-2023 Richard Rodger and other contributors, MIT License. */
 'use strict';
 var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
@@ -9,7 +9,7 @@ const Events = require('events');
 const Util = require('util');
 // External modules.
 const GateExecutor = require('gate-executor');
-const Jsonic = require('jsonic');
+const Jsonic = require('@jsonic/jsonic-next');
 const UsePlugin = require('use-plugin');
 const nid_1 = __importDefault(require("nid"));
 const patrun_1 = require("patrun");
@@ -30,14 +30,15 @@ const Add = require('./lib/add');
 const Sub = require('./lib/sub');
 const prior_1 = require("./lib/prior");
 const plugin_1 = require("./lib/plugin");
-const { Inward } = require('./lib/inward');
-const { Outward } = require('./lib/outward');
+const inward_1 = require("./lib/inward");
+const outward_1 = require("./lib/outward");
 const { Legacy } = require('./lib/legacy');
 const { resolve_options } = require('./lib/options');
 const Package = require('./package.json');
 const { Print } = require('./lib/print');
 const { addActions } = require('./lib/actions');
 const { transport } = require('./lib/transport');
+const package_json_1 = __importDefault(require("./package.json"));
 // Internal data and utilities.
 const { error, deep } = Common;
 // Seneca options.
@@ -84,7 +85,9 @@ const option_defaults = {
             callback: true,
             // Capture errors in actions and pass to callback (false throws uncaught).
             action: true,
-        }
+        },
+        // Custom function to identify thrown errors.
+        identify: (e) => e instanceof Error,
     },
     // Validate messages and options.
     valid: {
@@ -284,8 +287,8 @@ const option_defaults = {
 };
 // Utility functions exposed by Seneca via `seneca.util`.
 const seneca_util = {
-    Eraro: Eraro,
-    Jsonic: Jsonic,
+    Eraro,
+    Jsonic,
     Nid: nid_1.default,
     Patrun: patrun_1.Patrun,
     Gex: patrun_1.Gex,
@@ -371,32 +374,32 @@ function init(seneca_options, more_options) {
     });
     return seneca;
 }
-module.exports = init;
 // Expose Seneca prototype for easier monkey-patching
-module.exports.Seneca = Seneca;
-exports.default = init;
+init.Seneca = Seneca;
 // To reference builtin loggers when defining logging options.
-module.exports.loghandler = Legacy.loghandler;
+init.loghandler = Legacy.loghandler;
 // Makes require('seneca').use(...) work by creating an on-the-fly instance.
-module.exports.use = function top_use() {
+init.use = function top_use() {
     var argsarr = new Array(arguments.length);
     for (var l = 0; l < argsarr.length; ++l) {
         argsarr[l] = arguments[l];
     }
-    var instance = module.exports();
+    var instance = init();
     return instance.use.apply(instance, argsarr);
 };
 // Makes require('seneca').test() work.
-module.exports.test = function top_test() {
-    return module.exports().test(...arguments);
+init.test = function top_test() {
+    return init().test(...arguments);
 };
 // Makes require('seneca').quiet() work.
-module.exports.quiet = function top_quiet() {
-    return module.exports().quiet(...arguments);
+init.quiet = function top_quiet() {
+    return init().quiet(...arguments);
 };
-module.exports.util = seneca_util;
-module.exports.valid = Gubu;
-module.exports.test$ = { intern: intern };
+init.util = seneca_util;
+init.valid = Gubu;
+init.test$ = { intern: intern };
+exports.default = init;
+module.exports = init;
 // Create a new Seneca instance.
 function make_seneca(initial_opts) {
     // Create a private context.
@@ -416,7 +419,9 @@ function make_seneca(initial_opts) {
     };
     // These need to come from options as required during construction.
     private$.actrouter = start_opts.internal.actrouter || (0, patrun_1.Patrun)({ gex: true });
-    var soi_subrouter = start_opts.internal.subrouter || {};
+    private$.translationrouter =
+        start_opts.internal.translationrouter || (0, patrun_1.Patrun)({ gex: true });
+    const soi_subrouter = start_opts.internal.subrouter || {};
     private$.subrouter = {
         // Check for legacy inward router
         inward: soi_subrouter.inward || (0, patrun_1.Patrun)({ gex: true }),
@@ -438,9 +443,10 @@ function make_seneca(initial_opts) {
     // Define public member variables.
     root$.start_time = Date.now();
     root$.context = {};
-    root$.version = Package.version;
+    root$.version = package_json_1.default.version;
     // TODO: rename in 4.x as "args" terminology is legacy
     root$.fixedargs = {};
+    root$.fixedmeta = {};
     root$.flags = {
         closed: false,
     };
@@ -598,6 +604,7 @@ function make_seneca(initial_opts) {
         name: 'add',
         debug: !!start_opts.debug.ordu || !!start_opts.order.add.debug,
     })
+        .add(Add.task.translate)
         .add(Add.task.prepare)
         .add(Add.task.plugin)
         .add(Add.task.callpoint)
@@ -611,33 +618,33 @@ function make_seneca(initial_opts) {
         name: 'inward',
         debug: !!start_opts.debug.ordu || !!start_opts.order.inward.debug,
     })
-        .add(Inward.inward_msg_modify)
-        .add(Inward.inward_closed)
-        .add(Inward.inward_act_cache)
-        .add(Inward.inward_act_default)
-        .add(Inward.inward_act_not_found)
-        .add(Inward.inward_act_stats)
-        .add(Inward.inward_validate_msg)
-        .add(Inward.inward_warnings)
-        .add(Inward.inward_msg_meta)
-        .add(Inward.inward_limit_msg)
-        .add(Inward.inward_prepare_delegate)
-        .add(Inward.inward_sub)
-        .add(Inward.inward_announce);
+        .add(inward_1.Inward.inward_msg_modify)
+        .add(inward_1.Inward.inward_closed)
+        .add(inward_1.Inward.inward_act_cache)
+        .add(inward_1.Inward.inward_act_default)
+        .add(inward_1.Inward.inward_act_not_found)
+        .add(inward_1.Inward.inward_act_stats)
+        .add(inward_1.Inward.inward_validate_msg)
+        .add(inward_1.Inward.inward_warnings)
+        .add(inward_1.Inward.inward_msg_meta)
+        .add(inward_1.Inward.inward_limit_msg)
+        .add(inward_1.Inward.inward_prepare_delegate)
+        .add(inward_1.Inward.inward_sub)
+        .add(inward_1.Inward.inward_announce);
     root$.order.outward = new Ordu({
         name: 'outward',
         debug: !!start_opts.debug.ordu || !!start_opts.order.outward.debug,
     })
-        .add(Outward.outward_make_error)
-        .add(Outward.outward_act_stats)
-        .add(Outward.outward_act_cache)
-        .add(Outward.outward_res_object)
-        .add(Outward.outward_res_entity)
-        .add(Outward.outward_msg_meta)
-        .add(Outward.outward_trace)
-        .add(Outward.outward_sub)
-        .add(Outward.outward_announce)
-        .add(Outward.outward_act_error);
+        .add(outward_1.Outward.outward_make_error)
+        .add(outward_1.Outward.outward_act_stats)
+        .add(outward_1.Outward.outward_act_cache)
+        .add(outward_1.Outward.outward_res_object)
+        .add(outward_1.Outward.outward_res_entity)
+        .add(outward_1.Outward.outward_msg_meta)
+        .add(outward_1.Outward.outward_trace)
+        .add(outward_1.Outward.outward_sub)
+        .add(outward_1.Outward.outward_announce)
+        .add(outward_1.Outward.outward_act_error);
     // Configure logging
     // Mark logger as being externally defined from options
     if (start_opts.logger && 'object' === typeof start_opts.logger) {
